Update contributing-to-the-design-system.md
I suggested a few changes to this. I also am still confused after reading this what the "rules" are for moving an experimental component to the official design system. After a team validates it once, then can it move over? And if so, who decides that and who moves it over?