Update loading-indicator.md
Moved last bullet point under Accessibility considerations to Content considerations. Removed bullet point under Content considerations that said loading indicator could be used with or without a text label. According to feedback on #vagov-accessibility: 

"My gut reaction it should always have the text label -- and that the text label should have a status role, which functions as an aria live region for screen reader users. The text indicates when something is loading, and then updates and announces when loading is complete. Another thing to consider: Having a static (not animated) indicator for users with the prefers-reduced-motion flag enabled."

@humancompanion-usds and @pdavies88 can I merge this now? Any other changes?