Update index.md to add information about CAIA
CAIA wants to help VFS teams "shift left" and consider a11y, content, and IA issues as early as possible in their development processes. One way we feel we can do this is to draw more awareness to the existence of CAIA, what we do, and how teams can contact us for help.

We'd like to add a snippet of information about CAIA to the Accessibility page on design.va.gov to raise awareness and encourage early communications.

I added a section about getting guidance from the CAIA team on the bottom of this page.