Update index.md
As part of OCTO's initiative to improve accessibility documentation across VA.gov, this page has been restructured to  a) give a more thorough explanation of our approach, b) provide resources to non-VA teams, who use the VADS as a North Star, and c) give VA teams a more comprehensive starting point.

We're also working on a more internal-facing page that will be posted to Platform. We'll link to it here once that's complete.

If you have any questions, please reach out to Martha Wilkes. :) Thank you!